requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
aiosqlite>=0.17.0
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-multipart>=0.0.5
pillow>=8.3.1
aiohttp>=3.8.1
python-dotenv>=0.19.0
alembic>=1.7.4
openai>=1.3.0
aiofiles>=0.8.0
jsonschema>=4.17.3
greenlet>=1.1.2 